perf(www): isolate pricing page client graph
