+++ apps/www/app/[locale]/(app)/(shared)/(site)/page.tsx
@@ -10,7 +10,7 @@ import { Faq } from "@/components/marketing/about/faq/landing";
 import { Features } from "@/components/marketing/about/features";
 import { Hero } from "@/components/marketing/about/hero";
 import { Logos } from "@/components/marketing/about/logos";
-import { Pricing } from "@/components/marketing/about/pricing/plans";
+import { Pricing } from "@/components/marketing/about/pricing/landing";
 import { Schools } from "@/components/marketing/about/schools";
 import { Trust } from "@/components/marketing/about/trust/section";
 import {
